Validate when constructing Certificate model type
